requirements.txt
Django
pillow
gunicorn